size_t doesn't exist in xinu...

+++ Assignment5/xinunets/network/icmp/icmpResolve.c
@@ -3,7 +3,6 @@
  * @provides imcpResolve
  */
 
-#include <stdint.h>
 #include <xinu.h>
 #include "icmp.h"
 #include "../arp/arp.h"
@@ -74,7 +73,7 @@ process echoRequest(int dev, uchar* ipaddr, struct icmp_t* ping)
 
 int icmpResolve(uchar* ipaddr, struct icmp_t* ping)
 {
-	size_t icmp_size = (sizeof(struct ipgram) + sizeof(struct icmp_header_t));
+	int icmp_size = (sizeof(struct ipgram) + sizeof(struct icmp_header_t));
 	struct icmp_t* my_ping = (struct icmp_t*) malloc(icmp_size);
 	memcpy(my_ping, 0, icmp_size); // Clear out icmp packet
